Editorial: Use MV of Hex4Digits rather than SV
In general, SV returns a sequence of code units.
So, on the face of it, it's odd to require that
the result of SV fall within a range of integers.

Instead, it's simpler to phrase the requirement
in terms of the MV of |Hex4Digits|,
avoiding the detour though SV.