Editorial: more precise parameter for PrepareForOrdinaryCall (#3138)
In step 10 of the abstarct operation PrepareForOrdinaryCall, F.[[PrivateEnvironment]]
is accessed. This access is safe for an ECMAScript function object, as its
internal slots are mentioned in https://tc39.es/ecma262/#table-internal-slots-of-ecmascript-function-objects.
But in the case it is a built-in function object, this is not the case.
Fortunately, the only call sites of PrepareForOrdinaryCall are where the
argument is an ECMAScript function object, so we can make the parameter
 more precise.